Make responses trigger when the current user name appears in chat text.
At the moment, the bot is hardcoded to respond when it detects the string
"Derpy" in a text channel. This is suboptimal, as server administrators may
wish to change the name to something else. To improve usability, make
responses trigger on detection of the bot's displayed user name instead.

This will fix #7.